+++ .gitattributes
@@ -1,5 +1,2 @@
-/Tests export-ignore
 /.gitattributes export-ignore
 /.gitignore export-ignore
-/.travis.yml export-ignore
-/phpunit.xml.dist export-ignore
