+++ .gitignore
@@ -1,3 +1,4 @@
+/.idea/
 composer.lock
 vendor/*
 /phpunit.xml
