+++ composer.json
@@ -5,6 +5,12 @@
     "keywords": ["tree", "behaviors", "doctrine2", "extensions", "gedmo", "sluggable","loggable", "translatable", "nestedset", "sortable", "timestampable"],
     "homepage": "https://github.com/doctrine-extensions/DoctrineExtensionsSymfonyBundle",
     "license": "MIT",
+    "support": {
+        "issues": "https://github.com/doctrine-extensions/symfony-bundle/issues"
+    },
+    "config": {
+        "sort-packages": true
+    },
     "authors": [
         {
             "name": "Christophe Coevoet",
@@ -16,20 +22,25 @@
         }
     ],
     "require": {
-        "php": ">=5.3.2",
-        "symfony/framework-bundle": "~2.7|~3.2|~4.0",
+        "php": ">=7.1",
+        "symfony/framework-bundle": "^4.3.7 || ^5.0",
         "gedmo/doctrine-extensions": "^2.3.4"
     },
     "require-dev": {
         "roave/security-advisories": "dev-master",
-        "symfony/phpunit-bridge": "^4.0",
-        "symfony/security-bundle": "^2.7 || ^3.2 || ^4.0"
+        "symfony/phpunit-bridge": "^4.3.7 || ^5.0",
+        "symfony/security-bundle": "^4.3.7 || ^5.0"
     },
     "suggest": {
         "doctrine/doctrine-bundle": "to use the ORM extensions",
         "doctrine/mongodb-odm-bundle": "to use the MongoDB ODM extensions"
     },
+    "conflict": {
+        "stof/doctrine-extensions-bundle": "*"
+    },
     "autoload": {
-        "psr-4": { "DoctrineExtensions\\SymfonyBundle\\": "" }
+        "psr-4": {
+            "DoctrineExtensions\\SymfonyBundle\\": "src/"
+        }
     }
 }
