+++ src/DependencyInjection/Compiler/ValidateExtensionConfigurationPass.php
