+++ src/DependencyInjection/DoctrineExtensionsSymfonyBundleExtension.php
@@ -21,7 +21,7 @@ class DoctrineExtensionsSymfonyBundleExtension extends Extension
 
         $config = $processor->processConfiguration($configuration, $configs);
 
-        $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
 
         $loaded = array();
