+++ src/EventListener/BlameListener.php
@@ -2,28 +2,29 @@
 
 namespace DoctrineExtensions\SymfonyBundle\EventListener;
 
-use Gedmo\Loggable\LoggableListener;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
+use Gedmo\Blameable\BlameableListener;
+
 /**
  * Sets the username from the security context by listening on kernel.request
  *
- * @author Christophe Coevoet <stof@notk.org>
+ * @author David Buchmann <mail@davidbu.ch>
  */
-class LoggerListener implements EventSubscriberInterface
+class BlameListener implements EventSubscriberInterface
 {
     private $authorizationChecker;
     private $tokenStorage;
-    private $loggableListener;
+    private $blameableListener;
 
-    public function __construct(LoggableListener $loggableListener, TokenStorageInterface $tokenStorage = null, AuthorizationCheckerInterface $authorizationChecker = null)
+    public function __construct(BlameableListener $blameableListener, TokenStorageInterface $tokenStorage = null, AuthorizationCheckerInterface $authorizationChecker = null)
     {
-        $this->loggableListener = $loggableListener;
+        $this->blameableListener = $blameableListener;
         $this->tokenStorage = $tokenStorage;
         $this->authorizationChecker = $authorizationChecker;
     }
@@ -31,7 +32,7 @@ class LoggerListener implements EventSubscriberInterface
     /**
      * @internal
      */
-    public function onKernelRequest(GetResponseEvent $event)
+    public function onKernelRequest(RequestEvent $event)
     {
         if (HttpKernelInterface::MASTER_REQUEST !== $event->getRequestType()) {
             return;
@@ -42,9 +43,8 @@ class LoggerListener implements EventSubscriberInterface
         }
 
         $token = $this->tokenStorage->getToken();
-
         if (null !== $token && $this->authorizationChecker->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
-            $this->loggableListener->setUsername($token);
+            $this->blameableListener->setUserValue($token->getUser());
         }
     }
