+++ src/EventListener/LoggerListener.php
@@ -2,29 +2,28 @@
 
 namespace DoctrineExtensions\SymfonyBundle\EventListener;
 
+use Gedmo\Loggable\LoggableListener;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
-use Gedmo\Blameable\BlameableListener;
-
 /**
  * Sets the username from the security context by listening on kernel.request
  *
- * @author David Buchmann <mail@davidbu.ch>
+ * @author Christophe Coevoet <stof@notk.org>
  */
-class BlameListener implements EventSubscriberInterface
+class LoggerListener implements EventSubscriberInterface
 {
     private $authorizationChecker;
     private $tokenStorage;
-    private $blameableListener;
+    private $loggableListener;
 
-    public function __construct(BlameableListener $blameableListener, TokenStorageInterface $tokenStorage = null, AuthorizationCheckerInterface $authorizationChecker = null)
+    public function __construct(LoggableListener $loggableListener, TokenStorageInterface $tokenStorage = null, AuthorizationCheckerInterface $authorizationChecker = null)
     {
-        $this->blameableListener = $blameableListener;
+        $this->loggableListener = $loggableListener;
         $this->tokenStorage = $tokenStorage;
         $this->authorizationChecker = $authorizationChecker;
     }
@@ -32,7 +31,7 @@ class BlameListener implements EventSubscriberInterface
     /**
      * @internal
      */
-    public function onKernelRequest(GetResponseEvent $event)
+    public function onKernelRequest(RequestEvent $event)
     {
         if (HttpKernelInterface::MASTER_REQUEST !== $event->getRequestType()) {
             return;
@@ -43,8 +42,9 @@ class BlameListener implements EventSubscriberInterface
         }
 
         $token = $this->tokenStorage->getToken();
+
         if (null !== $token && $this->authorizationChecker->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
-            $this->blameableListener->setUserValue($token->getUser());
+            $this->loggableListener->setUsername($token);
         }
     }
