+++ src/Resources/config/sluggable.xml
