+++ src/Resources/config/softdeleteable.xml
