+++ src/Resources/doc/configuration.rst
