+++ src/Resources/doc/installation.rst
