+++ src/Resources/doc/uploadable-extension.rst
