+++ src/Uploadable/MimeTypeGuesserAdapter.php
